Simplified the interface for substitution models likelihood computation. Improved computation efficiency ...

+++ src/gui/Source/Character Matrices/RbData.h
@@ -28,7 +28,7 @@
 
 - (void)addTaxonData:(RbTaxonData*)td;
 - (void)addTaxonName:(NSString*)n;
-- (RbDataCell*)cellWithRow:(int)r andColumn:(int)c;
+- (RbDataCell*)cellWithRow:(size_t)r andColumn:(int)c;
 - (void)cleanName:(NSString*)nameStr;
 - (void)clear;
 - (RbData*)copiedFrom;
@@ -54,7 +54,7 @@
 - (void)restoreTaxonIndexed:(int)idx;
 - (void)restoreCharacterIndexed:(int)idx;
 - (void)setCopiedFrom:(RbData*)d;
-- (void)setNameOfTaxonWithIndex:(int)idx to:(NSString*)newName;
+- (void)setNameOfTaxonWithIndex:(size_t)idx to:(NSString*)newName;
 - (void)setStandardMatrixToHave:(int)rows andToHave:(int)columns;
 - (void)writeToFile:(NSString*)fn;
 - (NSString*)taxonWithIndex:(int)i;
