+++ src/gui/Source/Character Matrices/RbDataCell.h
@@ -12,20 +12,20 @@
 @interface RbDataCell : NSObject <NSCoding> {
 
 	BOOL                 isDiscrete;
-	int                  row;
-	int                  column;
+	size_t               row;
+	size_t               column;
 	int                  dataType;
-	int                  numStates;
+	size_t               numStates;
 	BOOL                 isAmbig;
     BOOL                 isGapState;
 	NSNumber*            val;
 }
 
 @property (readwrite)        BOOL      isDiscrete;
-@property (readwrite)        int       row;
-@property (readwrite)        int       column;
+@property (readwrite)        size_t    row;
+@property (readwrite)        size_t    column;
 @property (readwrite)        int       dataType;
-@property (readwrite)        int       numStates;
+@property (readwrite)        size_t    numStates;
 @property (readwrite)        BOOL      isAmbig;
 @property (readwrite)        BOOL      isGapState;
 @property (readwrite,retain) NSNumber* val;
